+++ packages/frontend/src/components/AccountListSidebar/AccountItem.tsx
@@ -7,6 +7,7 @@ import {
   onDCEvent,
   EffectfulBackendActions,
 } from '../../backend-com'
+import { runtime } from '@deltachat-desktop/runtime-interface'
 import { avatarInitial } from '../Avatar'
 import { getLogger } from '../../../../shared/logger'
 import useTranslationFunction from '../../hooks/useTranslationFunction'
@@ -133,7 +134,7 @@ export default function AccountItem({
     badgeContent = (
       <div
         className={classNames(styles.accountBadgeIcon, styles.bgSyncDisabled)}
-        aria-label='Background sync disabled'
+        aria-label={tx('background_sync_disabled_explaination')}
       >
         ⏻
       </div>
@@ -150,7 +151,7 @@ export default function AccountItem({
     )
   }
 
-  const ref = useRef<HTMLDivElement>(null)
+  const ref = useRef<HTMLButtonElement>(null)
   useLayoutEffect(() => {
     if (!isSelected) {
       return
@@ -186,7 +187,7 @@ export default function AccountItem({
   }, [isSelected, window.__screen])
 
   return (
-    <div
+    <button
       className={classNames(styles.account, {
         [styles.active]: isSelected,
         [styles['context-menu-active']]: isContextMenuActive,
@@ -204,7 +205,7 @@ export default function AccountItem({
           {account.profileImage ? (
             <img
               className={styles.content}
-              src={'file://' + account.profileImage}
+              src={runtime.transformBlobURL(account.profileImage)}
             />
           ) : (
             <div
@@ -232,7 +233,7 @@ export default function AccountItem({
         </div>
       )}
       <div className={classNames(styles.accountBadge)}>{badgeContent}</div>
-    </div>
+    </button>
   )
 }